+++ product_code/authentication_service/api/src/main/scala/de/upb/cs/uc4/authentication/model/AuthenticationUser.scala
@@ -20,6 +20,7 @@ case class AuthenticationUser(username: String, password: String, role: Authenti
 
   def validate(implicit ec: ExecutionContext): Future[Seq[SimpleError]] = Future {
     val usernameRegex = RegexCollection.AuthenticationUser.usernameRegex
+    val passwordRegex = RegexCollection.AuthenticationUser.passwordRegex
 
     var errors = List[SimpleError]()
     if (!usernameRegex.matches(username)) {
@@ -28,7 +29,7 @@ case class AuthenticationUser(username: String, password: String, role: Authenti
         "Username must consist of 4 to 16 characters, and must only contain letters, numbers, '-', and '.'."
       )
     }
-    if (password == "") {
+    if (!passwordRegex.matches(password)) {
       errors :+= SimpleError("password", "Password must not be empty.")
     }
     errors