+++ product_code/authentication_service/impl/src/main/scala/de/upb/cs/uc4/authentication/impl/AuthenticationApplication.scala
@@ -15,9 +15,12 @@ import de.upb.cs.uc4.authentication.impl.actor.{ AuthenticationBehaviour, Authen
 import de.upb.cs.uc4.authentication.impl.commands.DeleteAuthentication
 import de.upb.cs.uc4.authentication.impl.readside.{ AuthenticationDatabase, AuthenticationEventProcessor }
 import de.upb.cs.uc4.authentication.model.JsonUsername
+import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
+import de.upb.cs.uc4.shared.server.kafka.KafkaEncryptionComponent
 import de.upb.cs.uc4.shared.server.messages.Confirmation
 import de.upb.cs.uc4.shared.server.{ Hashing, UC4Application }
 import de.upb.cs.uc4.user.api.UserService
+import org.slf4j.{ Logger, LoggerFactory }
 import play.api.db.HikariCPComponents
 
 import scala.concurrent.Future
@@ -28,7 +31,10 @@ abstract class AuthenticationApplication(context: LagomApplicationContext)
   with SlickPersistenceComponents
   with JdbcPersistenceComponents
   with HikariCPComponents
-  with LagomKafkaComponents {
+  with LagomKafkaComponents
+  with KafkaEncryptionComponent {
+
+  private final val log: Logger = LoggerFactory.getLogger("Shared")
 
   private implicit val timeout: Timeout = Timeout(5.seconds)
 
@@ -53,12 +59,22 @@ abstract class AuthenticationApplication(context: LagomApplicationContext)
   lazy val userService: UserService = serviceClient.implement[UserService]
 
   userService
-    .userDeletedTopic()
+    .userDeletionTopic()
     .subscribe
     .atLeastOnce(
-      Flow.fromFunction[JsonUsername, Future[Done]](json =>
-        clusterSharding.entityRefFor(AuthenticationState.typeKey, Hashing.sha256(json.username))
-          .ask[Confirmation](replyTo => DeleteAuthentication(replyTo)).map(_ => Done)).mapAsync(8)(done => done)
+      Flow.fromFunction[EncryptionContainer, Future[Done]] { container =>
+        try {
+          val jsonUsername = kafkaEncryptionUtility.decrypt[JsonUsername](container)
+          clusterSharding.entityRefFor(AuthenticationState.typeKey, Hashing.sha256(jsonUsername.username))
+            .ask[Confirmation](replyTo => DeleteAuthentication(replyTo)).map(_ => Done)
+        }
+        catch {
+          case throwable: Throwable =>
+            log.error("AuthenticationService received invalid topic message: {}", throwable.toString)
+            Future.successful(Done)
+        }
+      }
+        .mapAsync(8)(done => done)
     )
 }