+++ product_code/configuration_service/CHANGELOG.md
@@ -1,6 +1,7 @@
 # [WIP v.0.10.2](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.10.1...configuration-v0.10.2) (2020-XX-XX)
 ## Feature
 ## Refactor
+- Changed the validation regex object to include more regexes
 ## Bugfix
 - Added minimum size to gzip (512 Byte)